refactor(hjb): factor HJB Newton/Picard loop into backend-agnostic WeakFormHJBSolver (#1131)
Sub-step (i)+(ii) of the weak-form solver work.

(i) Extend the WeakFormDiscretization protocol with `dof_coordinates` (the points
where nodal fields live and the Hamiltonian is evaluated). FEMDiscretization
returns `basis.doflocs.T` (== mesh.p.T for P1); MeshlessGalerkinDiscretization
returns its collocation nodes.

(ii) Introduce WeakFormHJBSolver(BaseHJBSolver): the implicit-Euler time stepping,
Picard linearization, and (semismooth) Newton iteration, written purely on the
assembled operators (K, M, gradient-projection R_d) + `dof_coordinates` + the
Hamiltonian. Boundary conditions are three subclass hooks
(`_is_pure_neumann` / `_dirichlet_dofs_and_values` / `_apply_bc_to_system`).
HJBFEMSolver becomes a thin subclass supplying the mesh+Lagrange discretization
and the scikit-fem BC strategy.

Gradient recovery is unified on the mass-lumped L2 projection `G_d = diag(1/M_lumped) @ R_d`;
this is identical (same integrand, same quadrature) to the FEM solver's prior
`_compute_nodal_gradient`, so behavior is preserved.

No regression: HJBFEMSolver was `_scheme_family = GENERIC` before and remains so.
FEM integration tests pass (12/12); scheme/factory/duality/three-mode-API subset
passes (161 passed, 8 skipped). Ruff check + format clean.

Part of #1131 Phase 2. Next: MeshlessGalerkinHJBSolver/FPSolver + MESHLESS_GALERKIN
scheme registration.
